Lógica da Segmented.
Tudo é culpa do Bahia.
