+++ Capivara/Screen/MyEvents/MyEventsTab.swift
@@ -1,0 +1,15 @@
+//
+//  MyEventsTab.swift
+//  Capivara
+//
+//  Created by Thallis Sousa on 02/12/21.
+//
+
+import Foundation
+
+enum MyEventsTab: String, CaseIterable {
+    case all = "Todos"
+    case current = "Atuais"
+    case created = "Criados"
+    case participated = "Participados"
+}
