+++ Capivara/Storage/Repositories/Mocks/EventRepositoryMock.swift
@@ -12,7 +12,9 @@ class EventRepositoryMock: EventRepositoryProtocol {
     var allEvents = EventMock.events
     
     func fetchUserEvents(of user: User) async throws -> [Event] {
-        allEvents
+        allEvents[3].creator = user
+        
+        return allEvents 
     }
     
     func fetchHotEvents() async throws -> [Event] {
